packages/atp-core/pyproject.toml
[project]
name = "atp-core"
version = "1.0.0"
description = "Core library for ATP Agent Test Platform — protocol, models, loader, and utilities"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "ruamel-yaml>=0.18",
    "jsonschema>=4.26.0",
    "structlog>=25.5.0",
    "cryptography>=46.0.5",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "prometheus-client>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-anyio>=0.0.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "atp/protocol",
    "atp/core",
    "atp/loader",
    "atp/chaos",
    "atp/cost",
    "atp/scoring",
    "atp/statistics",
    "atp/streaming",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
